Move PlaceParser to main
