+++ src/test/scala/bouken/world/PlaceBuilderTest.scala
@@ -1,17 +1,17 @@
-package bouken
+package bouken.world
 
 import bouken.domain._
 import org.scalatest.{FreeSpec, Matchers}
 
 import scala.util.Try
 
-class PlaceBuilderTest extends FreeSpec with Matchers {
+class PlaceParserTest extends FreeSpec with Matchers {
 
-  "PlaceBuilder" - {
+  "PlaceParser" - {
     "should build a basic tile" - {
 
       "when given '.'" - {
-        val place = PlaceBuilder.parse(".")
+        val place = PlaceParser.parse(".")
         "creates a ground tile" in {
           place.tile shouldBe Ground
         }
@@ -27,28 +27,28 @@ class PlaceBuilderTest extends FreeSpec with Matchers {
       }
 
       "when given '#'" - {
-        val place = PlaceBuilder.parse("#")
+        val place = PlaceParser.parse("#")
         "creates a wall tile" in {
           place.tile shouldBe Wall
         }
       }
 
       "when given ':'" - {
-        val place = PlaceBuilder.parse(":")
+        val place = PlaceParser.parse(":")
         "creates a rough tile" in {
           place.tile shouldBe Rough
         }
       }
 
       "when given 'w'" - {
-        val place = PlaceBuilder.parse("w")
+        val place = PlaceParser.parse("w")
         "creates a water tile" in {
           place.tile shouldBe Water
         }
       }
 
       "when given 'e50'" - {
-        val place = PlaceBuilder.parse("e50")
+        val place = PlaceParser.parse("e50")
         "creates an exit tile" in {
           place.tile shouldBe a[Exit]
         }
@@ -58,7 +58,7 @@ class PlaceBuilderTest extends FreeSpec with Matchers {
       }
 
       "when given '0Swamp'" - {
-        val place = PlaceBuilder.parse("""/0Swamp""")
+        val place = PlaceParser.parse("""/0Swamp""")
         "creates a stair tile" in {
           place.tile shouldBe Stairs(To(0, "Swamp"))
         }
@@ -68,7 +68,7 @@ class PlaceBuilderTest extends FreeSpec with Matchers {
     "should build a place with an effect" - {
 
       "when given '.|,'" - {
-        val place = PlaceBuilder.parse(".|,")
+        val place = PlaceParser.parse(".|,")
         "creates a ground tile" in {
           place.tile shouldBe Ground
         }
@@ -84,7 +84,7 @@ class PlaceBuilderTest extends FreeSpec with Matchers {
       }
 
       "when given '.|;'" - {
-        val place = PlaceBuilder.parse(".|;")
+        val place = PlaceParser.parse(".|;")
         "creates a ground tile" in {
           place.tile shouldBe Ground
         }
@@ -94,7 +94,7 @@ class PlaceBuilderTest extends FreeSpec with Matchers {
       }
 
       "when given '.|+'" - {
-        val place = PlaceBuilder.parse(".|+")
+        val place = PlaceParser.parse(".|+")
         "creates a ground tile" in {
           place.tile shouldBe Ground
         }
@@ -104,7 +104,7 @@ class PlaceBuilderTest extends FreeSpec with Matchers {
       }
 
       "when given '.|g'" - {
-        val place = PlaceBuilder.parse(".|g")
+        val place = PlaceParser.parse(".|g")
         "creates a ground tile" in {
           place.tile shouldBe Ground
         }
@@ -117,7 +117,7 @@ class PlaceBuilderTest extends FreeSpec with Matchers {
     "should build a place with an occupier" - {
 
       "when given '.|Z'" - {
-        val place = PlaceBuilder.parse(".|Z")
+        val place = PlaceParser.parse(".|Z")
         "creates a ground tile" in {
           place.tile shouldBe Ground
         }
@@ -136,7 +136,7 @@ class PlaceBuilderTest extends FreeSpec with Matchers {
       }
 
       "when given '.|X'" - {
-        val place = PlaceBuilder.parse(".|X")
+        val place = PlaceParser.parse(".|X")
         "creates a tile occupied by an Enemy" in {
           place.state shouldBe a[Enemy]
         }
@@ -146,7 +146,7 @@ class PlaceBuilderTest extends FreeSpec with Matchers {
       }
 
       "when given '.|M'" - {
-        val place = PlaceBuilder.parse(".|M")
+        val place = PlaceParser.parse(".|M")
         "creates a tile occupied by an Enemy" in {
           place.state shouldBe a[Enemy]
         }
@@ -156,84 +156,4 @@ class PlaceBuilderTest extends FreeSpec with Matchers {
       }
     }
   }
-}
-
-object PlaceBuilder {
-
-  def parse(value: String): Place =
-    if (value.exists(_ == '|')) handleComplexPlace(value)
-    else makeTile(parseTile(value))
-
-  private def handleComplexPlace(value: String): Place = {
-    val tileStr = value.takeWhile(_ != '|')
-    val otherStr = value.dropWhile(_ != '|').tail
-
-    val tile = parseTile(tileStr)
-
-    val effect = otherStr match {
-      case "," => Some(Trap(2))
-      case ";" => Some(Snare(2))
-      case "+" => Some(Heal(2))
-      case "g" => Some(Gold(3))
-      case _   => None
-    }
-
-    val enemy = otherStr match {
-      case "Z" => Some(Enemy(Zombie))
-      case "X" => Some(Enemy(Gnoll))
-      case "M" => Some(Enemy(Minotaur))
-      case _   => None
-    }
-
-    (effect, enemy) match {
-      case (Some(e), _) => makeTile(tile, tileEffect = e)
-      case (_, Some(e)) => makeTile(tile, state = e)
-      case _ => makeTile(tile)
-    }
-  }
-
-  private def parseTile(string: String) =
-    string match {
-      case "." => Ground
-      case ":" => Rough
-      case "#" => Wall
-      case "w" => Water
-      case e if e.startsWith("e") && Try(e.drop(1).toInt).toOption.isDefined => Exit(Score(e.tail.toInt))
-      case s if stairsCheck(s) => makeStairs(s)
-      case _ => Ground
-    }
-
-  private def makeTile(
-    tile: Tile = Ground,
-    visible: Boolean = false,
-    state: Occupier = Empty,
-    tileEffect: TileEffect = NoEffect
-    ) = Place(
-      visible,
-      tile,
-      state,
-      tileEffect
-    )
-
-  private def stairsCheck(string: String): Boolean =
-    string.length >= 3 &&
-      string.charAt(0) == '/' &&
-      Try(string.charAt(1).toInt).toOption.isDefined
-
-  private def makeStairs(string: String): Stairs = {
-    val value = string.tail
-    val id = value.takeWhile(x => Numbers.contains(x)).toInt
-    val level = value.dropWhile(x => Numbers.contains(x))
-
-    Stairs(To(id, level))
-  }
-
-  private val Numbers = "0123456789"
-
-  private val EmptyTile = Place(
-    visible = false,
-    tile = Ground,
-    state = Empty,
-    tileEffect = NoEffect
-  )
 }
